added many supports

+++ main.py
@@ -3,9 +3,9 @@ import pickle
 import sys
 from tkinter import *
 from tkinter import messagebox
-
 import pyAesCrypt
 import pygame
+import os.path
 
 pygame.init()
 bufferSize = 64 * 1024
@@ -37,10 +37,10 @@ blue = (0, 0, 255)
 green = (0, 255, 0)
 catch_error = True
 
-fb_user_text = ''
+fb_user_text = ""
 active_fb = False
 
-font = pygame.font.Font('freesansbold.ttf', 30)
+font = pygame.font.Font("freesansbold.ttf", 30)
 
 
 def text_object(text, font, color):
@@ -52,30 +52,69 @@ def text_object(text, font, color):
 
 
 def message_display_small(text, a, b, color, display):
-    smalltext = pygame.font.Font('comic.ttf', 30)
+    smalltext = pygame.font.Font("comic.ttf", 30)
     textsurf, textrect = text_object(text, smalltext, color)
     textrect.center = (int(a), int(b))
     display.blit(textsurf, textrect)
 
 
 def fb_text(text, a, b, color, display):
-    smalltext = pygame.font.Font('freesansbold.ttf', 30)
+    smalltext = pygame.font.Font("freesansbold.ttf", 30)
     textsurf, textrect = text_object(text, smalltext, color)
     textrect.center = (int(a), int(b))
     display.blit(textsurf, textrect)
 
 
-def fb_button():
+def fb_button(username, password):
+    file_name = str(username) + str(password) + ".bin"
+    if os.path.exists(file_name):
+        root = Tk()
+        f1 = open(file_name, "rb")
+        line = pickle.load(f1)
+        root.title("Facebook Account")
+        first = line[0]
+        second = line[1]
+        a1_text = Label(root, text=first)
+        a2_text = Label(root, text=second)
+        a1_text.grid(row=0, column=0)
+        a2_text.grid(row=1, column=0)
+    else:
+        second = Tk()
+        second.title("Facebook Login")
+        username1 = Label(second, text="Username:")
+        password1 = Label(second, text="Password:")
+        username1_entry = Entry(second)
+        password1_entry = Entry(second, show="*")
+        login_text1 = "Please provide Facebook account and password"
+        change = "Do you want to change password or username"
+        change_button = Button(second, text=change)
+        login_text = Label(second, text=login_text1)
+        username1.grid(row=2, column=0)
+        password1.grid(row=3, column=0)
+        username1_entry.grid(row=2, column=1)
+        password1_entry.grid(row=3, column=1, columnspan=2)
+        username2 = username1_entry.get()
+        password2 = password1_entry.get()
+        username_list = []
 
-    root = Tk()
+        def save():
+            username_list.append(username2)
+            username_list.append(password2)
+            fie = str(username) + str(password)
+            f = open(fie + ".bin", "wb")
+            pickle.dump(username_list, f)
+            f.close()
+
+        saving = Button(second, text="Save", command=save)
+        saving.grid(row=3, column=1)
+        change_button.grid(row=4, column=1)
 
 
-def gameloop(a, username, passwor, file):
-    fb = 'Facebook'
-    update = True
+def gameloop(a, file):
+    fb = "Facebook"
+    quitting = True
     while True:
-        if update == True:
-            a.fill((255, 255, 255))
+        a.fill((255, 255, 255))
         for e in pygame.event.get():
             if e.type == pygame.QUIT:
                 pygame.quit()
@@ -85,15 +124,23 @@ def gameloop(a, username, passwor, file):
         mouse = pygame.mouse.get_pressed()
         mouse_pos = pygame.mouse.get_pos()
         a.blit(facebook, (20, 20))
-        message_display_small(fb, 45 + facebook.get_width() - 100 + 10,
-                              25 + facebook.get_height() + 10, black, a)
-        if mouse[0] == 1 and 20 < mouse_pos[0] < 20 + facebook.get_width() and 20 < mouse_pos[1] < 20 + facebook.get_height():
-            fb_button()
-            update = False
-        if update == True:
+        message_display_small(
+            fb,
+            45 + facebook.get_width() - 100 + 10,
+            25 + facebook.get_height() + 10,
+            black,
+            a,
+        )
+        if (
+            mouse[0] == 1
+            and 20 < mouse_pos[0] < 20 + facebook.get_width()
+            and 20 < mouse_pos[1] < 20 + facebook.get_height()
+        ):
+            pygame.quit()
+            quitting = False
+            fb_button(username, password)
+        if quitting:
             pygame.display.update()
-        else:
-            pass
 
 
 def login():
@@ -105,15 +152,19 @@ def login():
     lbl = Label(login_window, text="Please enter your username and password:")
 
     def check():
-        testing = True
+        testing = False
         password = pass_entry.get()
         username = input_entry.get()
         main_password = username + "" + password
         file_name = username + password
         try:
             pyAesCrypt.decryptFile(
-                file_name + '.bin.aes', file_name + 'decrypted' + '.bin', main_password, bufferSize)
-            f = open(file_name + 'decrypted' + '.bin', 'rb')
+                file_name + ".bin.aes",
+                file_name + "decrypted" + ".bin",
+                main_password,
+                bufferSize,
+            )
+            f = open(file_name + "decrypted" + ".bin", "rb")
             logins = pickle.load(f)
             for a in logins:
                 if a[1] == password:
@@ -122,19 +173,17 @@ def login():
                     messagebox.showinfo("Success", "Success")
                     login_window.destroy()
                     root.destroy()
-            testing = False
-        except:
             testing = True
+        except:
+            testing = False
             root = Tk()
             root.withdraw()
             messagebox.showinfo("Error", "Wrong Password or Username")
             root.destroy()
-        if testing == False:
+        if testing:
             d = pygame.display.set_mode((800, 600))
-            gameloop(d, username, password, file_name + 'decrypted' + '.bin')
+            gameloop(d, file_name + "decrypted" + ".bin")
 
-        else:
-            pass
     but = Button(login_window, text="Login", command=check)
     login.grid(row=2, column=2)
     lbl.grid(row=0, column=2, columnspan=2)
@@ -160,20 +209,27 @@ def register():
     a = []
 
     def inputing():
-        f = open("user.bin", "ab")
-        l = []
         password = pass_entry1.get()
         username = input_entry1.get()
-        l.append(username)
-        l.append(password)
-        a.append(l)
-        pickle.dump(a, f)
-        HSP = username + "" + password
-        f.close()
-        file_name = username + password
-        pyAesCrypt.encryptFile("user.bin", file_name +
-                               '.bin.aes', HSP, bufferSize)
-        os.remoce('user.bin')
+        if os.path.exists(username + password + ".bin"):
+            messagebox.showinfo("Error", "The account with the same username exist!!")
+        else:
+            f = open(username + password + ".bin", "wb")
+            l = []
+            l.append(username)
+            l.append(password)
+            a.append(l)
+            pickle.dump(a, f)
+            HSP = username + "" + password
+            f.close()
+            file_name = str(username) + str(password)
+            pyAesCrypt.encryptFile(
+                str(username) + str(password) + ".bin",
+                file_name + ".bin.aes",
+                HSP,
+                bufferSize,
+            )
+
     but = Button(login_window1, text="Register", command=inputing)
 
     lbl1 = Label(login_window1, text=text)
@@ -189,8 +245,7 @@ def register():
 
 main = Label(root, text="Welcome to ONE-PASS manager")
 login_text = Label(root, text="Do you already have an account")
-register_text = Label(
-    root, text='If you don"t have an account please register')
+register_text = Label(root, text='If you don"t have an account please register')
 reg_button = Button(root, text="Register", command=register)
 login_button = Button(root, text="login", command=login)  # added login button
